fix dashboard page
