+++ app/(protected)/dashboard/@dashboard/page.tsx
@@ -1,31 +1,11 @@
 'use client';
 
-// import DashboardContent from '@/components/DashboardContent';
-import { useProviderCallback } from '@/hooks/useAuth';
-import { useSearchParams } from 'next/navigation';
+import DashboardContent from '@/components/DashboardContent';
 
 export default function DashboardSlot() {
-  const searchParams = useSearchParams();
-
-  const code = searchParams.get('code') || '';
-  const scope = searchParams.get('scope') || '';
-  const authuser = searchParams.get('authuser') || '';
-  const prompt = searchParams.get('prompt') || '';
-
-  const { error, isLoading } = useProviderCallback('google', {
-    code,
-    scope,
-    authuser,
-    prompt,
-  });
-
   return (
-    <div className="p-4">
-      <h1 className="text-2xl font-bold">Dashboard</h1>
-
-      {isLoading && <p>Loading user data...</p>}
-      {error && <p className="text-red-500">Error: {error}</p>}
-      {/* <DashboardContent />; */}
+    <div>
+      <DashboardContent />
     </div>
   );
 }
