+++ app/(protected)/dashboard/@profile/page.tsx
@@ -1,18 +1,18 @@
 'use client';
 
-import { useAuth } from '@/hooks/useAuth';
+// import { useAuth } from '@/hooks/useAuth';
 
 export default function ProfileSlot() {
-  const { user, isLoading, error } = useAuth();
+  // const { user, isLoading, error } = useAuth();
 
-  if (isLoading) return <div>Loading...</div>;
-  if (error) return <div className="text-destructive">{error}</div>;
+  // if (isLoading) return <div>Loading...</div>;
+  // if (error) return <div className="text-destructive">{error}</div>;
 
   return (
     <div>
-      <h2 className="text-lg font-semibold">User Profile</h2>
+      {/* <h2 className="text-lg font-semibold">User Profile</h2>
       <p>Email: {user?.email}</p>
-      <p>Name: {user?.name}</p>
+      <p>Name: {user?.name}</p> */}
     </div>
   );
 }
