--- app/(protected)/overview/page.tsx
@@ -1,12 +1,0 @@
-import React from 'react';
-import OverviewComponent from '@/components/overView';
-
-function OverViewPage() {
-  return (
-    <>
-      <OverviewComponent />
-    </>
-  );
-}
-
-export default OverViewPage;
