+++ components/DashboardContent.tsx
@@ -1,18 +1,30 @@
 'use client';
 
-import { useAuth } from '@/hooks/useAuth';
+import { useProviderCallback } from '@/hooks/useAuth';
+import OverviewComponent from './overView';
+import { useSearchParams } from 'next/navigation';
 
 export default function DashboardContent() {
-  const { user, isLoading, error } = useAuth();
+  const searchParams = useSearchParams();
+
+  const code = searchParams.get('code') || '';
+  const scope = searchParams.get('scope') || '';
+  const authuser = searchParams.get('authuser') || '';
+  const prompt = searchParams.get('prompt') || '';
+
+  const { isLoading, error } = useProviderCallback('google', {
+    code,
+    scope,
+    authuser,
+    prompt,
+  });
 
   if (isLoading) return <div>Loading...</div>;
   if (error) return <div className="text-destructive">{error}</div>;
 
   return (
-    <div className="p-4">
-      <h1 className="text-2xl font-bold">Dashboard</h1>
-      <p>Welcome, {user?.name}!</p>
-      <p>Email: {user?.email}</p>
+    <div>
+      <OverviewComponent />
     </div>
   );
 }
